+++ changelog.md
@@ -25,6 +25,8 @@ Codion Change Log
 - AttributeDefinition.Builder.captionResourceKey() and mnemonicResourceKey() renamed to captionResource() and mnemonicResource().
 - AttributeDefinition.Builder.captionResource() and mnemonicResource() overloaded to use a specified resource bundle instead of the entity bundle.
 - Column.ColumnDefiner.column() parameter changed from Function to ColumnTemplate.
+### is.codion.framework.db
+- EntityConnectionProvider.description() now returns Optional.
 ### is.codion.framework.model
 - DefaultEntityTableConditionModel bug fixed, now handles column and dateTimeFormat precision mismatch, by using ranges when precision is lacking.
 ### is.codion.swing.framework.ui